% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_profiles.R
\name{load_profile_data}
\alias{load_profile_data}
\title{Load L2H profile data}
\usage{
load_profile_data(folder_path)
}
\arguments{
\item{folder_path}{Path to the folder of data files. These files should end in "txt"}
}
\value{
A tibble of the L2H profile data
}
\description{
Given a directory of L2H data files, this function collects all of the data
and stores it in a tibble. Unused channels are removed automatically.
}
